requirements.txt
appdirs==1.4.2
dj-config-url==0.1.1
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.4
geopy==1.11.0
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0
